Add files via upload
In this JavaScript project, we will make an image search engine where we can search for any image by typing a name and it will display a list of images from stock images website Unsplash. Here we are using Unsplash API to make this image search engine with the help of JavaScript.